requirements.txt
uWSGI==2.0.17
aniso8601==4.0.1
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
